[ADD] xml_data_generator: Created xml_data_generator module
[IMP] xml_data_generator: Added support for exporting data to xml

[IMP] xml_data_generator: Added order for xml records

[IMP] xml_data_generator: Added correct dependency arrangement

[REF] xml_data_generator: Attendend lints

[FIX] xml_data_generator: Finally fixed dependency order for sure

Additionally, refactored code to slightly improfe performance
and readability (still there is a lot of room for imprevement there).

Finally, added checks to prevent going too deep recursively, and
corrected the way duplicates are avoided so that they can appear in
evals and refs but not generated as xml records.

[REF] xml_data_generator: Replaced 'xml_id' by 'external_id' everywhere

[REF] xml_data_generator: Added 'auto' to new generated xml data

[IMP] xml_data_generator: Addded export wizard action to all models

Additionally, renamed files to comply with OCA standards and fixed lints.

[IMP] xml_data_generator: Changed 'recursive_depth' to a selection field

[MIG] xml_data_generator: Migrated to v15 from v16