[FIX] xml_data_generator: Finally fixed dependency order for sure
Additionally, refactored code to slightly improfe performance
and readability (still there is a lot of room for imprevement there).

Finally, added checks to prevent going too deep recursively, and
corrected the way duplicates are avoided so that they can appear in
evals and refs but not generated as xml records.